+++ WalletWasabi.Fluent/Models/Wallets/IWalletAuthModel.cs
@@ -1,0 +1,18 @@
+using System.Threading.Tasks;
+
+namespace WalletWasabi.Fluent.Models.Wallets;
+
+public interface IWalletAuthModel
+{
+	bool IsLoggedIn { get; }
+
+	bool IsLegalRequired { get; }
+
+	Task<WalletLoginResult> TryLoginAsync(string password);
+
+	Task AcceptTermsAndConditions();
+
+	void CompleteLogin();
+
+	void Logout();
+}
