+++ WalletWasabi.Fluent/Models/Wallets/IWalletLoadWorkflow.cs
@@ -1,0 +1,10 @@
+namespace WalletWasabi.Fluent.Models.Wallets;
+
+public interface IWalletLoadWorkflow
+{
+	bool IsLoading { get; }
+
+	void Start();
+
+	IObservable<(double PercentComplete, TimeSpan TimeRemaining)> Progress { get; }
+}
